requirements.txt
flask==3.0.3
python-dotenv==1.0.1
gunicorn==23.0.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
resend